requirements.txt
fastapi
uvicorn
requests
cryptography
pyodbc
python-dotenv
qrcode[pil]